requirements.txt
flask
flask_sqlalchemy
flask_migrate       # для Alembic
flask_jwt_extended  # если JWT
flask_cors
pydantic             # или marshmallow
python-dotenv
